# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/clientes": {
      "get": {
        "description": "Este endpoint sirve para obtener un cliente en específico",
        "operationId": "ClientesController_getAllClientes",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtener todos los clientes con paginación",
        "tags": [
          "Clientes"
        ]
      },
      "post": {
        "description": "Este endpoint sirve para crear un nuevo cliente",
        "operationId": "ClientesController_createCliente",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClienteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Para crear un nuevo cliente",
        "tags": [
          "Clientes"
        ]
      }
    },
    "/clientes/{idCliente}": {
      "get": {
        "description": "Este endpoint sirve para obtener un cliente en específico",
        "operationId": "ClientesController_getCliente",
        "parameters": [
          {
            "name": "idCliente",
            "required": true,
            "in": "path",
            "description": "ID del cliente a obtener",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtener un cliente por su id",
        "tags": [
          "Clientes"
        ]
      },
      "patch": {
        "description": "Este endpoint sirve para poder actualizar un cliente",
        "operationId": "ClientesController_updateCliente",
        "parameters": [
          {
            "name": "idCliente",
            "required": true,
            "in": "path",
            "description": "ID del cliente a actualizar",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClienteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Actualizar un cliente existente",
        "tags": [
          "Clientes"
        ]
      }
    },
    "/clientes/{idCliente}/delete": {
      "patch": {
        "description": "Este endpoint permite eliminar un cliente cambiando su estado a inactivo",
        "operationId": "ClientesController_deleteCliente",
        "parameters": [
          {
            "name": "idCliente",
            "required": true,
            "in": "path",
            "description": "ID del cliente a eliminar",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Eliminar un cliente a elección",
        "tags": [
          "Clientes"
        ]
      }
    },
...
    "/departamentos": {
      "get": {
        "description": "Este endpoint lista todos los departamentos activos con paginación",
        "operationId": "DepartamentosController_getAllDepartamentos",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtener todos los departamentos",
        "tags": [
          "Departamentos"
        ]
      }
    },
    "/departamentos/{idDepartamento}": {
      "get": {
        "description": "Este endpoint sirve para obtener un departamento específico por su ID",
        "operationId": "DepartamentosController_getDepartamento",
        "parameters": [
          {
            "name": "idDepartamento",
            "required": true,
            "in": "path",
            "description": "ID del departamento a obtener",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtener un departamento por ID",
        "tags": [
          "Departamentos"
        ]
      }
    },
...
    "/lotes": {
      "get": {
        "operationId": "LotesController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lotes obtenida correctamente"
          }
        },
        "summary": "Obtener todos los lotes activos con paginación",
        "tags": [
          "lotes"
        ]
      },
      "post": {
        "operationId": "LotesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lote creado correctamente"
          },
          "400": {
            "description": "Datos inválidos para creación"
          }
        },
        "summary": "Crear un nuevo lote",
        "tags": [
          "lotes"
        ]
      }
    },
    "/lotes/{id}": {
      "get": {
        "operationId": "LotesController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del lote",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lote encontrado exitosamente"
          },
          "404": {
            "description": "Lote no encontrado o está desactivado"
          }
        },
        "summary": "Obtener un lote por ID",
        "tags": [
          "lotes"
        ]
      },
      "patch": {
        "operationId": "LotesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLoteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lote actualizado exitosamente"
          },
          "400": {
            "description": "Datos inválidos para actualización"
          },
          "404": {
            "description": "Lote no encontrado para actualizar"
          }
        },
        "summary": "Actualizar un lote existente",
        "tags": [
          "lotes"
        ]
      }
    },
    "/lotes/{id}/delete": {
      "patch": {
        "operationId": "LotesController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lote desactivado correctamente"
          },
          "400": {
            "description": "El lote tiene stock activo y no puede desactivarse"
          },
          "404": {
            "description": "Lote no encontrado o ya desactivado"
          }
        },
        "summary": "Desactivar un lote (borrado lógico)",
        "tags": [
          "lotes"
        ]
      }
    },
...
    "/auth/login": {
      "post": {
        "description": "Autentica al usuario y devuelve tokens JWT y refresh token",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Iniciar sesión",
        "tags": [
          "Autenticación"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Genera nuevos tokens de acceso usando un refresh token válido",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Refrescar tokens",
        "tags": [
          "Autenticación"
        ]
      }
    }
  },
  "info": {
    "title": "Control de Endpoints",
    "description": "Documentación de la API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateClienteDto": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "description": "Nombre del cliente",
            "example": "Juan Pérez"
          },
          "telefonoCliente": {
            "type": "string",
            "description": "Teléfono del cliente",
            "example": "1234567890"
          },
          "correoCliente": {
            "type": "string",
            "description": "Correo electrónico del cliente",
            "example": "example@gmail.com"
          },
          "nitCliente": {
            "type": "string",
            "description": "NIT del cliente",
            "example": "12345678"
          },
          "idMunicipio": {
            "type": "number",
            "description": "Se ingresa el id del municipio al que pertenece el cliente",
            "example": 1
          }
        },
        "required": [
          "nombreCliente"
        ]
      },
      "UpdateClienteDto": {
        "type": "object",
        "properties": {
          "nombreCliente": {
            "type": "string",
            "description": "Nombre del cliente",
            "example": "Juan Pérez"
          },
          "telefonoCliente": {
            "type": "string",
            "description": "Teléfono del cliente",
            "example": "1234567890"
          },
          "correoCliente": {
            "type": "string",
            "description": "Correo electrónico del cliente",
            "example": "exampleUpdated@gmail.com"
          },
          "nitCliente": {
            "type": "string",
            "description": "NIT del cliente",
            "example": "1234567890"
          }
        }
      },
...
      "CreateLoteDto": {
        "type": "object",
        "properties": {
          "fechaVencimiento": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de vencimiento del lote",
            "example": "2024-12-31"
          },
          "numeroLote": {
            "type": "string",
            "description": "Número único del lote",
            "example": "LOTE-2023-001"
          }
        },
        "required": [
          "fechaVencimiento",
          "numeroLote"
        ]
      },
      "UpdateLoteDto": {
        "type": "object",
        "properties": {
          "fechaVencimiento": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de vencimiento actualizada",
            "example": "2025-01-15"
          },
          "numeroLote": {
            "type": "string",
            "description": "Número de lote actualizado",
            "example": "LOTE-2023-001-REV"
          }
        }
      },
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "identificador": {
            "type": "string",
            "example": "juan123 o juan@ejemplo.com"
          },
          "contrasena": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "identificador",
          "contrasena"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "Refresh token para obtener nuevo access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refresh_token"
        ]
      }
    }
  }
}